further refinement of plugins
